Fixed code review comments and svc registrations
