+++ src/test/activation/downloader.unit.test.ts
@@ -9,8 +9,7 @@ import { expect } from 'chai';
 import { SemVer } from 'semver';
 import * as TypeMoq from 'typemoq';
 import { LanguageServerDownloader } from '../../client/activation/downloader';
-import { IPlatformData } from '../../client/activation/platformData';
-import { ILanguageServerFolderService } from '../../client/activation/types';
+import { ILanguageServerFolderService, IPlatformData } from '../../client/activation/types';
 import { IApplicationShell } from '../../client/common/application/types';
 import { IFileSystem } from '../../client/common/platform/types';
 import { IOutputChannel } from '../../client/common/types';
